Develop (#60)
* issue_55: change exception format

* iss 5, 55, 29, 40 (51), 44, 48

* temporary changes

* Issue_5: nested set using nestedj

* fix fb

* issue_5: finished impl

* issue_5: cleanup code, add comments

* issue_44: change request param to request body

* bugfix/fix view endpoints

* issue_48: relatives

* issue_40: aadd swagger conf

* issue_51: change User to UserPrincipal

* issue_40: ignore auth principal

* issue_5: fix mistake

* issue_5: fix field name

* issue_48: fix sorting

* issue_48: fix dtos

* issue_40: exclude error handler, add doc description

* issue_29: push service

* issue_5: update nestedj to 4.1.0, fix removing

* add privacy policy

* issue_35: add de locale

* issue_59: camel case

* issue_35: some references translation

* issue_35: update doc

* issue_59: upd doc

* bugfix/long id

* wrap responces to data

* fix migration

* cors

* add user delete endpoint

* rename users endpoint

* sort fix

* createdAt

* published field

* bugfix

* fix nested set

* float fix, add near filter validation, fix nested

* migration fix

* dont allow watch nonpublished posts

+++ src/main/java/ru/smartel/strike/dto/request/event/EventFiltersDTO.java
@@ -70,23 +70,23 @@ public class EventFiltersDTO extends PostFiltersDTO {
     }
 
     public static class SearchArea {
-        private Double lat;
-        private Double lng;
+        private Float lat;
+        private Float lng;
         private Integer radius;
 
-        public Double getLat() {
+        public Float getLat() {
             return lat;
         }
 
-        public void setLat(Double lat) {
+        public void setLat(Float lat) {
             this.lat = lat;
         }
 
-        public Double getLng() {
+        public Float getLng() {
             return lng;
         }
 
-        public void setLng(Double lng) {
+        public void setLng(Float lng) {
             this.lng = lng;
         }
