+++ src/main/java/ru/smartel/strike/specification/conflict/NearCoordinateConflict.java
@@ -15,11 +15,11 @@ import javax.persistence.criteria.Root;
  * https://stackoverflow.com/questions/21084886/how-to-calculate-distance-using-latitude-and-longitude
  */
 public class NearCoordinateConflict implements Specification<Conflict> {
-    private Double latitude;
-    private Double longitude;
+    private Float latitude;
+    private Float longitude;
     private Integer radius;
 
-    public NearCoordinateConflict(Double latitude, Double longitude, Integer radius) {
+    public NearCoordinateConflict(Float latitude, Float longitude, Integer radius) {
         this.latitude = latitude;
         this.longitude = longitude;
         this.radius = radius;
